Fix for when a field name is null
I've customized a RepeaterMatrix field in certain ProcessWire templates to remove certain fields when accessed from those templates. I believe this to be the problem. Rather than checking which fields are available and only looping through them, this simply runs a check to verify that the field name exists prior to entering the `foreach` loop in the Indexer's ___getRepeatableIndexValue() method.

Although further testing is recommended, this particular addition should not cause any problems, and only prevents an edge case error.